+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add backend for multi-slack support for mobile-app @Ferril ([#3500](https://github.com/grafana/oncall/pull/3500))
+
 ## v1.3.72 (2023-12-05)
 
 ### Fixed